refhash.t:  Quiet warnings
Originally, the program needed to test whether, for several sets of two
variables each, the definedness of one variable matched the definedness
of the second or (if that condition was not met) the two variables were
string-equal.  Anticipating that, if warnings were generally on, these
statements would spew "uninitialized value" warnings, the program
silenced them with:

        local $^W = 0;

However, when warnings are on by default, that statement is ineffective.
We can get more precise results by scoping "no warnings
'uninitialized';" to the block of statements likely to warn.