t/re/subst.t:  Adapt to honor strict-by-default
Avoid use of $a and $b, as they are special to 'sort'.  Once these are renamed
throughout, we can run with strictures except in 3 blocks where we were
previously saying 'local *a'.  In those 3 blocks, we relax strict vars.

Re-order BEGIN block to be consistent with strictures and other test files in this
directory.